Convert double into float in the whole project

+++ ray.h
@@ -7,8 +7,8 @@
 struct Ray {
 	Vec pos, dir;
 	Ray(Vec _p = Vec(), Vec _d = Vec());
-	bool intersect_with_sphere(const Sphere& sphere, double& dist) const;
-	bool intersect_with_triangle(const Triangle& tri, double& dist) const;
+	bool intersect_with_sphere(const Sphere& sphere, float& dist) const;
+	bool intersect_with_triangle(const Triangle& tri, float& dist) const;
 	bool intersect_with_box(const Vec& low, const Vec& high) const;
 };
