+++ vec.h
@@ -3,16 +3,16 @@
 
 class Vec {
 public:
-	double x, y, z;
-	Vec(double _x = 0.0, double _y = 0.0, double _z = 0.0);
-	double length() const;
+	float x, y, z;
+	Vec(float _x = 0.0, float _y = 0.0, float _z = 0.0);
+	float length() const;
 	Vec normal() const;
 	Vec blend(const Vec& rhs) const;
 	Vec operator + (const Vec& rhs) const;
 	Vec operator - () const;
 	Vec operator - (const Vec& rhs) const;
-	double operator * (const Vec& rhs) const;
-	Vec operator * (const double rhs) const;
+	float operator * (const Vec& rhs) const;
+	Vec operator * (const float rhs) const;
 	Vec operator % (const Vec& rhs) const;
 };
